backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.5.2
psutil==6.1.0
python-dotenv==1.0.1
docker==7.1.0
huawei-lte-api==1.9.3
requests==2.32.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
aiofiles==24.1.0
pytz==2024.2
tzdata==2024.2
